# apps/api/src/authn/auth.controller.ts
import { Body, Controller, Get, HttpCode, Inject, Post, Req, Res } from '@nestjs/common';
import type { Request, Response } from 'express';
import {
  LoginRequest,
  RegisterRequest,
  type LoginRequestDto,
  type MeResponseDto,
  type RegisterRequestDto,
} from '@duckoj/contracts';
import { AppError } from '../common/app.error.js';
import { ZodValidationPipe } from '../common/zod.pipe.js';
import { APP_CONFIG } from '../config/config.module.js';
import type { AppConfig } from '../config/config.schema.js';
import type { Actor } from '../authz/actor.js';
import { AuthService, toMe } from './auth.service.js';
import { SessionService } from './session.service.js';
import { TotpService } from './totp.service.js';
import { CurrentActor, Public } from './auth.guard.js';

@Controller('auth')
export class AuthController {
  constructor(
    @Inject(AuthService) private readonly auth: AuthService,
    @Inject(SessionService) private readonly sessions: SessionService,
    @Inject(TotpService) private readonly totp: TotpService,
    @Inject(APP_CONFIG) private readonly config: AppConfig,
  ) {}

  @Post('register')
  @Public()
  @HttpCode(201)
  register(
    @Body(new ZodValidationPipe(RegisterRequest)) body: RegisterRequestDto,
  ): Promise<MeResponseDto> {
    return this.auth.register(body);
  }

  @Post('login')
  @Public()
  @HttpCode(200)
  async login(
    @Body(new ZodValidationPipe(LoginRequest)) body: LoginRequestDto,
    @Req() req: Request,
    @Res({ passthrough: true }) res: Response,
  ): Promise<{ user: MeResponseDto }> {
    const user = await this.auth.login(body.usernameOrEmail, body.password);
    const totpEnabled = await this.totp.isEnabled(user.id);
    if (totpEnabled) {
      if (!body.totpCode) {
        throw new AppError(401, 'totp_required', 'A two-factor code is required.');
      }
      if (!(await this.totp.verify(user.id, body.totpCode))) {
        throw new AppError(401, 'invalid_totp_code', 'That code is not valid.');
      }
    }
    const { token, expiresAt } = await this.sessions.issue(user.id, {
      ip: req.ip,
      userAgent: req.get('user-agent') ?? undefined,
    });
    res.cookie(this.config.sessionCookieName, token, {
      httpOnly: true,
      sameSite: 'lax',
      secure: this.config.nodeEnv === 'production',
      path: '/',
      expires: expiresAt,
    });
    return { user: toMe(user, totpEnabled) };
  }

  // Public because logging out is idempotent: a caller whose session has
  // already expired should still get its cookie cleared, not a 401.
  @Post('logout')
  @Public()
  @HttpCode(204)
  async logout(@Req() req: Request, @Res({ passthrough: true }) res: Response): Promise<void> {
    const token = req.cookies?.[this.config.sessionCookieName] as string | undefined;
    if (token) await this.sessions.revoke(token);
    res.clearCookie(this.config.sessionCookieName, { path: '/' });
  }

  @Get('me')
  async me(@CurrentActor() actor: Actor): Promise<MeResponseDto> {
    const user = await this.auth.loadUser(actor.userId);
    return toMe(user, await this.totp.isEnabled(user.id));
  }
}